refactor: SessionRepository 분리
