+++ src/main/java/nextstep/courses/domain/session/SessionRepository.java
@@ -1,0 +1,7 @@
+package nextstep.courses.domain.session;
+
+public interface SessionRepository {
+    void save(Long courseId, Session session);
+
+    Sessions findByCourseId(Long courseId);
+}
